chore(app): move webview centric app into its own package
