+++ app/webviewcentric/proguard-rules.pro
