+++ app/webviewcentric/src/androidTest/java/com/swiften/templateapp/webviewcentric/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.swiften.templateapp
+package com.swiften.templateapp.webviewcentric
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
