+++ app/webviewcentric/src/main/AndroidManifest.xml
@@ -5,12 +5,12 @@
     android:allowBackup="true"
     android:icon="@mipmap/ic_launcher"
     android:label="@string/app_name"
-    android:name="com.swiften.templateapp.MainApplication"
+    android:name=".MainApplication"
     android:roundIcon="@mipmap/ic_launcher_round"
     android:supportsRtl="true"
     android:theme="@style/Theme.TemplateApp">
     <activity
-      android:name="com.swiften.templateapp.MainActivity"
+      android:name=".MainActivity"
       android:exported="true">
       <intent-filter>
         <action android:name="android.intent.action.MAIN"/>
