+++ app/webviewcentric/src/main/java/com/swiften/templateapp/webviewcentric/MainFragment.kt
@@ -1,4 +1,4 @@
-package com.swiften.templateapp
+package com.swiften.templateapp.webviewcentric
 
 import android.content.SharedPreferences
 import android.os.Bundle
@@ -7,8 +7,8 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.google.gson.Gson
-import com.swiften.templateapp.databinding.MainFragmentBinding
-import com.swiften.templateapp.webview.AppJavascriptInterface
+import com.swiften.templateapp.webviewcentric.databinding.MainFragmentBinding
+import com.swiften.templateapp.webviewcentric.webview.AppJavascriptInterface
 import com.swiften.webview.BridgeMethodArgumentsParser
 import com.swiften.webview.BridgeRequestProcessor
 import com.swiften.webview.SharedPreferencesJavascriptInterface
