+++ app/webviewcentric/src/main/java/com/swiften/templateapp/webviewcentric/Redux.kt
@@ -1,4 +1,4 @@
-package com.swiften.templateapp
+package com.swiften.templateapp.webviewcentric
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
