+++ app/webviewcentric/src/main/java/com/swiften/templateapp/webviewcentric/utils.kt
@@ -1,4 +1,4 @@
-package com.swiften.templateapp
+package com.swiften.templateapp.webviewcentric
 
 fun getLogTag(target: Any): String {
   return target.javaClass.simpleName
