+++ app/webviewcentric/src/main/res/values/strings.xml
